Alunos corretos por atividade

+++ src/main/java/br/com/unincor/webSite/model/dao/AlunoDao.java
@@ -3,7 +3,7 @@ package br.com.unincor.webSite.model.dao;
 import br.com.unincor.webSite.model.domain.Aluno;
 import jakarta.persistence.Query;
 
-public class AlunoDao extends GenericDao<Aluno, Long> {
+public class AlunoDao extends GenericDao<Aluno, Integer> {
     
      public Aluno buscarAlunoPorLogin(String login) {
         String sql = "from Aluno a where a.email = :email";
